requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
astroid==2.11.7
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
cryptography==37.0.4
cssselect==1.1.0
decorator==5.1.1
dicttoxml==1.7.4
dill==0.3.5.1
Django==4.1
django-crontab==0.7.1
django-ipware==4.0.2
django-qr-code==3.1.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
docutils==0.19
frozenlist==1.3.1
future==0.18.2
geocoder==1.38.1
geoip2==4.6.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
ip2geotools==0.1.6
IP2Location==8.8.0
isodate==0.6.1
isort==5.10.1
jeepney==0.8.0
keyring==23.7.0
lazy-object-proxy==1.7.1
lxml==4.9.1
maxminddb==2.2.0
mccabe==0.7.0
multidict==6.0.2
outcome==1.2.0
packaging==21.3
pip-review==1.1.1
pkginfo==1.8.3
platformdirs==2.5.2
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pylint==2.14.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyquery==1.4.3
PySocks==1.7.1
pytz==2022.1
qrcode==7.3.1
ratelim==0.1.6
readme-renderer==36.0
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
SecretStorage==3.3.2
segno==1.5.2
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
twine==4.0.1
typed-ast==1.5.4
typing==3.7.4.3
typing_extensions==4.3.0
urllib3==1.26.11
webencodings==0.5.1
wrapt==1.14.1
wsproto==1.1.0
yarl==1.8.1
zeep==4.1.0
zipp==3.8.1
